refactor(synapsys): extract budget-aware renderer to separate module
Move the budget-aware renderer cluster (decideInjection policy, reverse-walk
demotion, ledger commit, cortex_query-augmented body formatting, budget
constants) from hooks/synapsys.js to lib/render-budget.js to comply with the
static-gate 400-line file budget. Dispatcher is now 277 lines, render-budget.js
206 lines. Call sites and signatures preserved; behavior unchanged.
