+++ plugins/synapsys/lib/render-budget.js
@@ -1,0 +1,206 @@
+'use strict';
+
+/**
+ * Budget-aware memory renderer (extracted from hooks/synapsys.js to keep the
+ * dispatcher under the static-gate line budget).
+ *
+ * This module owns the per-memory injection policy (decideInjection), the
+ * reverse-walk budget demotion (renderMatchedMemories → demoteToFit), the
+ * ledger commit semantics, and the cortex_query-augmented body formatting.
+ * The dispatcher consumes `renderMatchedMemories`, `formatMemoryForRender`,
+ * and the shared `SEP` separator. Behavior is byte-for-byte identical to the
+ * pre-extraction inline implementation.
+ */
+
+const path = require('node:path');
+const injectLedger = require(path.join(__dirname, 'inject-ledger'));
+const { demoteToFit } = require(path.join(__dirname, 'budget'));
+const { appendCortexQuery } = require(path.join(__dirname, 'cortex-hook'));
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Budget constants (brief P0 R1 / R3 / spec §P0 #1).
+//
+// MAX_INJECT_CHARS — soft cap on total injected text. Memories that cause the
+//   matched set to exceed this limit are demoted to summary form (reverse-walk),
+//   never silently dropped (brief P0 R8 / spec §P0 #8).
+// SKIP_DEMOTION_BELOW — memories whose full body is below this size are never
+//   chosen for demotion: their full text is small enough to always inject
+//   in full (brief P0 R3 / spec §P0 #3).
+//
+// Both may be overridden at runtime via `SYNAPSYS_INJECT_BUDGET` (positive
+// integer; brief P2 R12 / spec §P2 #1). See `resolveActiveBudget`.
+// ─────────────────────────────────────────────────────────────────────────────
+const MAX_INJECT_CHARS = 16000;
+const SKIP_DEMOTION_BELOW = 2000;
+const SEP = '\n\n---\n\n';
+
+function resolveActiveBudget() {
+  const raw = process.env.SYNAPSYS_INJECT_BUDGET;
+  if (raw == null || raw === '') return MAX_INJECT_CHARS;
+  const parsed = parseInt(raw, 10);
+  return Number.isFinite(parsed) && parsed > 0 ? parsed : MAX_INJECT_CHARS;
+}
+
+/**
+ * decideInjection — pure helper implementing the brief AC-5 renderer policy.
+ *
+ *   always       → full body on every match
+ *   once         → full body iff injectedCount === 0, else reminder
+ *   occasionally → full body iff injectedCount % fireCadence === 0, else reminder
+ *
+ * `ledgerEntry` is the per-memory record from `loadLedger().memories[name]`
+ * (or `undefined` for "never injected this session").
+ */
+function resolveCadence(memory) {
+  const raw = Number(memory && memory.fireCadence);
+  return raw > 0 ? raw : 5;
+}
+
+function decideInjection(memory, ledgerEntry) {
+  const mode = (memory && memory.fireMode) || 'once';
+  const count = Number(ledgerEntry && ledgerEntry.injectedCount) || 0;
+  if (mode === 'always') return { kind: 'full' };
+  if (mode === 'occasionally') {
+    const cadence = resolveCadence(memory);
+    return { kind: count % cadence === 0 ? 'full' : 'reminder' };
+  }
+  // default: once
+  return { kind: count === 0 ? 'full' : 'reminder' };
+}
+
+function reminderLine(memory) {
+  return `[synapsys:active] ${memory.name} (fired earlier; full body in this session)`;
+}
+
+function formatMemory(m) {
+  const header = `[synapsys:${m.store.kind}] ${m.name} — ${m.description || '(no description)'}`;
+  if (m.inject === 'full') {
+    return `${header}\n\n${m.body.trim()}`;
+  }
+  return `${header}\n  source: ${m.file}`;
+}
+
+/**
+ * Render a memory's full body, augmented with its Phase 2 cortex_query block
+ * when one applies. `cortexCtx` is built once per dispatch by
+ * `cortexQueryContext`; a null ctx (e.g. fail-open fallback paths) yields the
+ * plain body. This is the body fed into the budget-aware renderer so cortex
+ * recall output is governed by the same injection budget as memory text.
+ */
+function formatMemoryForRender(memory, cortexCtx) {
+  const base = formatMemory(memory);
+  if (!cortexCtx || !cortexCtx.recall) return base;
+  return appendCortexQuery(base, memory, cortexCtx);
+}
+
+function commitInjection(ledger, sessionId, memory, isFull) {
+  const entry = ledger.memories[memory.name];
+  const prevCount = Number(entry && entry.injectedCount) || 0;
+  const prevLast = Number(entry && entry.lastFullInjectAt) || 0;
+  const nextCount = prevCount + 1;
+  ledger.memories[memory.name] = {
+    injectedCount: nextCount,
+    lastFullInjectAt: isFull ? nextCount : prevLast,
+  };
+  try {
+    injectLedger.recordInjection(sessionId, memory.name, { full: isFull });
+  } catch {
+    /* fail-open */
+  }
+}
+
+// Budget-aware renderer (brief P0 R1/R2/R4–R8). After the per-memory
+// decideInjection pass, run a reverse-walk demotion to bring the total under
+// `activeBudget`. Ledger semantics (brief P0 R6):
+//   initialKind='full'  && finalKind='full'     → commitInjection(..., true)
+//   initialKind='reminder'                       → commitInjection(..., false)
+//   initialKind='full'  && finalKind='reminder' → NO commitInjection (re-fires
+//                                                  in full on the next match).
+// The whole call is wrapped in `try` so any throw falls open to the plain
+// formatMemory join — memory injection must never block the user (spec §Security).
+function buildEntry(memory, ledgerMemories, cortexCtx) {
+  const kind = decideInjection(memory, ledgerMemories[memory.name]).kind;
+  return {
+    memory,
+    initialKind: kind,
+    finalKind: kind,
+    fullText: formatMemoryForRender(memory, cortexCtx),
+    summaryText: reminderLine(memory),
+  };
+}
+
+function emitEntries(entries, ledger, sessionId) {
+  let demotedCount = 0;
+  const pieces = [];
+  for (const e of entries) {
+    const isFull = e.finalKind === 'full';
+    pieces.push(isFull ? e.fullText : e.summaryText);
+    if (e.initialKind === 'full' && e.finalKind === 'reminder') {
+      // Budget-induced demotion: do NOT bump the ledger so the memory
+      // re-fires in full on the next match (brief P0 R6 / G5).
+      demotedCount += 1;
+      continue;
+    }
+    commitInjection(ledger, sessionId, e.memory, isFull);
+  }
+  return { body: pieces.join(SEP), demotedCount };
+}
+
+function writeStderrLine(line) {
+  try {
+    process.stderr.write(line);
+  } catch {
+    /* fail-open */
+  }
+}
+
+function emitBudgetAlerts(demotedCount, bodyLength, activeBudget) {
+  // Stderr alert (brief P0 R7 / spec §Security: count-only, no names/bodies).
+  if (demotedCount > 0) {
+    writeStderrLine(
+      `[synapsys] ${demotedCount} memories summarized to fit ${activeBudget}-char budget — they will inject in full on next match.\n`
+    );
+  }
+  // Debug stderr line when SYNAPSYS_DEBUG=1 (brief P1 R11).
+  if (process.env.SYNAPSYS_DEBUG === '1') {
+    writeStderrLine(`[synapsys:debug] budget ${bodyLength}/${activeBudget}\n`);
+  }
+}
+
+/**
+ * renderMatchedMemories — per-memory loop wrapper. Routes each match through
+ * the ledger + decideInjection + recordInjection. The entire call is fail-open
+ * (R1): any throw → fall back to formatting every memory as full body.
+ */
+function renderMatchedMemories(matched, sessionId, cortexCtx) {
+  try {
+    const ledger = injectLedger.loadLedger(sessionId);
+    if (!ledger.memories || typeof ledger.memories !== 'object') {
+      ledger.memories = {};
+    }
+    const activeBudget = resolveActiveBudget();
+    const entries = matched.map((m) => buildEntry(m, ledger.memories, cortexCtx));
+    demoteToFit(entries, {
+      limit: activeBudget,
+      sep: SEP,
+      skipBelow: SKIP_DEMOTION_BELOW,
+    });
+    const { body, demotedCount } = emitEntries(entries, ledger, sessionId);
+    emitBudgetAlerts(demotedCount, body.length, activeBudget);
+    return body;
+  } catch {
+    return matched.map((m) => formatMemoryForRender(m, null)).join(SEP);
+  }
+}
+
+module.exports = {
+  SEP,
+  MAX_INJECT_CHARS,
+  SKIP_DEMOTION_BELOW,
+  resolveActiveBudget,
+  resolveCadence,
+  decideInjection,
+  formatMemory,
+  formatMemoryForRender,
+  renderMatchedMemories,
+};
